requirements.txt
# Async and Network
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
anyio==4.8.0
httpcore==1.0.7
httpx==0.28.1
h11==0.14.0
idna==3.10
sniffio==1.3.1
urllib3==2.3.0

# Web Framework
Flask==3.1.0
Flask-SocketIO==5.5.1
Werkzeug==3.1.3
python-socketio==5.12.1
python-engineio==4.11.2
itsdangerous==2.2.0
click==8.1.8
Jinja2==3.1.5
simple-websocket==1.1.0
wsproto==1.2.0

# AI and Language Models
langchain==0.3.17
langchain-core==0.3.33
langchain-openai==0.3.3
langchain-text-splitters==0.3.5
langsmith==0.3.3
openai==1.60.2
openai-whisper==20240930
tiktoken==0.8.0

# Data Processing
numpy==2.1.3
scipy==1.15.1
torch==2.6.0
networkx==3.4.2
sympy==1.13.1
numba==0.61.0
llvmlite==0.44.0

# Audio Processing
PyAudio==0.2.14
sounddevice==0.5.1
pydub==0.25.1
pyttsx3==2.98
webrtcvad==2.0.10

# Azure Services
azure-cognitiveservices-speech==1.42.0
azure-core==1.32.0
azure-ai-vision-imageanalysis==1.0.0


# Data Validation
pydantic==2.10.6
pydantic_core==2.27.2
jsonpatch==1.33
jsonpointer==3.0.0
annotated-types==0.7.0

# Utils
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
setuptools==75.8.0
packaging==24.2
more-itertools==10.6.0
six==1.17.0
regex==2024.11.6
tqdm==4.67.1
typing_extensions==4.12.2
PyYAML==6.0.2

# Windows Specific
pypiwin32==223
pywin32==308
comtypes==1.4.9

# Misc
attrs==25.1.0
bidict==0.23.1
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
distro==1.9.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2024.12.0
greenlet==3.1.1
jiter==0.8.2
MarkupSafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
orjson==3.10.15
pycparser==2.22
pygame==2.6.1
propcache==0.2.1
SQLAlchemy==2.0.37
tenacity==9.0.0
yarl==1.18.3
zstandard==0.23.0
